+++ src/content/panorama/scripts/custom_game/game_setup/game_setup.js
@@ -5383,7 +5383,7 @@ function loadPlayerBans() {
         if (!preloadedHeroPanels[data.heroName]) {
             var heroImage = $.CreatePanel('Panel', $.GetContextPanel(), 'reviewPhaseHeroImageLoader');
 
-            heroImage.BLoadLayoutFromString('<root><Panel><DOTAScenePanel style="width: 300px; height: 800px; opacity-mask: url(\'s2r://panorama/images/masks/softedge_box_png.vtex\');" unit="' + data.heroName + '"/></Panel></root>', false, false);
+            heroImage.BLoadLayoutFromString('<root><Panel><DOTAScenePanel particleonly="false" style="width: 300px; height: 800px; opacity-mask: url(\'s2r://panorama/images/masks/softedge_box_png.vtex\');" unit="' + data.heroName + '"/></Panel></root>', false, false);
             heroImage.AddClass("avatarScene");    
 
             heroImage.visible = false;